[deb-x64, rpm-x64] Force umask to 0022 within containers. (#11)
The latest rvm update (1.29.9) introduced a change in the shell
script preparing the rvm environment at login. As a result, the
umask in the container has changed, which changes the permissions
of all files created within the containers. Because of this, the
file included in the deb and rpm packages had their permissions
changed, which can lead to installation issues.