[deb-armhf] Compile ruby from source (#72)
Forces compilation of ruby from source in the 32-bit ARM image.

Inside the 32-bit ARM image, the kernel arch is advertised as aarch64 (as the host is an ARM64 machine). Thus rvm downloads the precompiled aarch64 binary of ruby 2.3.8 (which got added on Aug. 29), making the use of ruby fail due to incompatible archs.