[windows] Update msys distribution (#80)
Updates the msys distribution to the latest available: packages for previous distributions have been pruned from the msys repositories (see https://www.msys2.org/news/).

Uses ridk install 2 to force-update to the latest packages. The packages provided in the latest msys release are broken (cmake fails to use the cc compiler), so we have to get the fixed packages by upgrading to the latest available packages.