Add additional checking to the aws-networkign script. (#119)
The set-network route appears to throw an error, but the route is
successfully applied.  The failure leaves a message in the build output
indicating it failed (when it didn't).

Apply the change, and if/when the command returns an error, check that
the resulting route is properly set, and write the result accordingly.
If the route is successful, then ignore the error.