[windows] Pin bundler version to 2.3.22 (#298)
bundler 2.3.23 was released yesterday. It breaks the Windows build: if given the same Gemfile in datadog-agent as bundler 2.3.22, it doesn't install the win32-process gem, which breaks the Agent build.

To fix this, this PR pins bundler to 2.3.22 in the Windows images.

Long-term, we should consider systematically pinning bundler in all places where it's used. It's the second time a patch update of bundler breaks things.