Install GCC 10.4.0
This change installs GCC 10.4.0 on the following build images:

 - deb-arm
 - deb-x64
 - rpm-arm64
 - rpm-x64
 - suse-x64

We chose to provide our own GCC version, because the GCC versions
available as part of CentOS 6, openSUSE 42.1 and Ubuntu 14.04
weren't recent enough to build OpenSCAP and some of its dependencies.

We chose GCC 10.4.0, because it was the most recent version that
could be built using the GCC versions available on our distributions.
GCC 11 and upper include libcody, which is written in C++11
and requires a more recent GCC version to be built.

This change is the result of a collaborative work between
Sylvain Baubeau, Pierre Guilleminot and David du Colombier.

Co-authored-by: Sylvain Baubeau <lebauce@gmail.com>
Co-authored-by: Pierre Guilleminot <pierre.guilleminot@datadoghq.com>