Update the python install script.
When installing on developer machine (not in build container) noticed
the unexpected side-effect that various python requirements couldn't
be installed due to lack of write permissions.

In the developer install case, install a virtual environment.  Update
the powershell module to load that virtual environment so that the
developer mode is always working on that private environment