[windows] Update container root certificates (#391)
There is a possible timing issue in the test and build scripts in the datadog-agent repository: the base container doesn't contain some of the root certificates needed to install dependencies. These certificates are added while the container is running (probably by a Windows Update process), but if this happens too late, the test and build Windows jobs can fail.

To prevent this, we force the certificates update to happen while the container is being built.

Long-term version of this short-term fix: https://github.com/DataDog/datadog-agent/pull/17207, which allows to not query the new certificates on every test / build, making it flaky (due to possible network issues).