Update install_java.ps1 to use absolute path for java.zip
When running the install script on a VM, `java.zip` was not downloaded to the correct directory (it was downloaded to `C:\Windows\System32\`). Update the script to download the .zip file to the same directory that the script is executing in and use `$out` to extract the file content.