circleci: install clang format from the usual ubuntu repos
This repo contains version which are much more bleeding edge than the
one we use for the agent build.
As new option appear, we are hitting a situation where the circleci
version wants to apply a change that the deb-x86 images want to revert
This leads to a situation where we can't have a green CI.
An alternative solution would be to find the setting that causes the
change and attempt to set it to a value which makes both versions happy,
but there's no guaranty of this being possible.