[ABLD-135] Add `bazelisk` to Windows containers (#953)
* Add `bazelisk` to Windows containers

As the title implies, this is to install the `bazelisk` command line
tool (standalone binary distribution) meant to bootstrap `bazel`, since
the version of `bazel` itself will be branch-pinned for reproducible
builds.

The change does its best to stick to helpers used in sibling
installation scripts.

* Make `bazel.exe` a symlink to `bazelisk.exe`

As per review comment:
https://github.com/DataDog/datadog-agent-buildimages/pull/953#discussion_r2283162334