feat(build.sh): Use two-layered cache approach
We include the branch name in the cache key for our images.
Then, if we are running on a PR branch, the build can pull the cache from both the branch-related cache (i.e. the build cache from previous commits on the PR), falling back to main if there is no hit.
When pushing to the cache, we only push to the branch-related cache. Note that this way the `main`-related cache only gets updated once the `main` commit build runs, i.e. when the PR is merged.