[ACIX-1069] Fix docker image caching (#993)
* fix(build.sh): Include architecture info in cache key

* feat(build.sh): Use two-layered cache approach

We include the branch name in the cache key for our images.
Then, if we are running on a PR branch, the build can pull the cache from both the branch-related cache (i.e. the build cache from previous commits on the PR), falling back to main if there is no hit.
When pushing to the cache, we only push to the branch-related cache. Note that this way the `main`-related cache only gets updated once the `main` commit build runs, i.e. when the PR is merged.

* feat(build.sh): Add `set -x` for easier debugging

* fix(build.sh): Sanitize branch names for docker tag format

* [TEMP] Modify Dockerfile to check cache

* Revert "[TEMP] Modify Dockerfile to check cache"

This reverts commit 53523d2d01af9f7dcd856278e3a560568a74fb16.

* fix(build): Prevent trailing slash at the end of sanitized cache keys

`echo` (or maybe the pipe ?) adds a newline character which would get transformed by `tr` into an underscore.

* fix(build.sh): Refer to official docs

* fix(build.sh): Add comment explaining `\n` in `sanitize`