fix: Use file-specific bind mounts in `RUN` directives
This has a pretty big impact on the cache invalidation logic. When we mount the entire repo, like was done before, the cached layer gets invalidated when there is _any_ change to _any_ file in the repo, no matter if it actually impacts the layer or not.
By only mounting the individual script file, we make sure that only changes to the script file affect the build.