feat(linux): Make datadog-packages installation work locally and in CI
The problem solved here is that we were using a private repository for datadog-packages, so we need to authenticate to the repo in order to download it.
In CI, this is easy enough to do with the CI_JOB_TOKEN environment variable, but locally we need to use a "real" GitLab token.
This logic is now expressed flexibly in the bakefile.