[ACIX-1056] Fix bind mounts in docker builds (#1034)
### What does this PR do?

Replaces all `RUN --mount` instructions in the Dockerfiles to specifically mount the files it needs instead of the entire repo.

Also moves a bunch of logic from the dockerfile itself into separate scripts for better readability and maintainability.

### Motivation

Mounting the entire repo _destroys_ caching performance: with the entire repo mounted, the cached value of each layer is invalidated as soon as _any_ file in the repo gets changed.

By mounting only the necessary file(s) in each `RUN` layer, we make sure that these layer's cached values are invalidated if and only if these specific files get changed, instead of any file in the repo.

### Possible Drawbacks / Trade-offs

None, as long as every image builds properly. Maybe it's a bit more verbose.

### Additional Notes

See #958.
I did not make the change for the Windows image as it requires mounting a lot more files - I will probably do it in a follow-up PR. 


Tested in https://github.com/DataDog/datadog-agent/pull/41927