[ACIX-1056] Make `linux/Dockerfile` multistage (#1038)
* refactor(linux): Start use of same format as builtin `BUILDARCH`

Standardizing on one format for arch definition, and only specifying aliases when useful will make it a lot easier to read and implement. Using the builtin `BUILDARCH` also avoids us needing to pass in an extra build arg for the build platform.

* feat(linux): Split Dockerfile into multiple stages

This allows for better layer caching and smaller final image size.
This also accelerates the build process as all stages are built in parallel.

* refactor(linux): Move build args closer to their usage

This allows for improved caching performance, as changing a build arg now only invalidates the stage that uses it, instead of the entire base stage.

* fix(linux): Build `texinfo` and `crosstool-ng` together

For some reason crosstool-ng does not build properly without texinfo being built first. Seems to be some issue with the libc install that gets resolved by building texinfo first.

* feat(linux): Use multiple base images for more parallelization

* fix(linux): Add missing apt update commands

* fix(linux): AI-suggested fixes for missing things

* fix(build.sh): Don't pass build arguments with empty values

Default-valued ARGs in Dockerfiles don't use the default value if a build arg is passed with an empty value. (i.e. passing `--build-arg BASE_IMAGE=` in a Dockerfile with `BASE_IMAGE=ubuntu:24.04` will result in an empty `BASE_IMAGE`)

* fix: fix datadog_packages_builder stage not finding go

* fix(build.sh): Add args from `go.env` and `dda.env` separately

* fix: Add missing ARG instructions to multistage dockerfile

* fix(linux): Use `build_base` for bazelisk (needs curl)

* fix(linux): Make `conda init bash` work properly

This cryptic instruction was present in the old Dockerfile and has now been restored in the proper places

* fix(linux): Add missing rust version arg from final stage

Since RUST_VERSION is not passed as a build arg the build relies on the default value. Without one we end up running `rustup default` with no version specified, which is not what we want.

* refactor(linux): Remove LABEL steps

The env vars containing the versions don't carry over between stages anyway so these would all be empty.
With Bake we will be able to make this work as expected.

* refactor(linux): Copilot-suggested cleanup

* fix(linux): Fix AWS CLI installation

* fix(linux): Move `CONDA_PATH` arg

* fix(linux): Add missing automake package

This is necessary for building the dev envs.

* fix(linux): More missing things from final stage

* fix(linux): Include recommended packages in `apt install`

Seems we were relying on some auto-installed recommended packages in the previous image (like `semodule-utils`). We should probably be more explicit about what packages we require, and avoid relying on packages that are only installed as deps of other packages - but this can be done at a later time.

* fix(linux): Include msgo toolchain in final image

* refactor(build.sh): Move reused code to function

* chore(linux): Remove redundant comments