[ACIX-1056] Use Docker Bake for `linux` image (#1039)
* Move setup scripts to `linux/scripts`

* Restore main-branch script to setup/

* feat(linux): Create docker bake file regrouping all build args

feat(linux): Simplify up go, python and bazel install scripts for multiarch

We move a lot of the logic from the shell script for these tools directly into the dockerfile/bakefile:
- Having the versions and checksums in the bakefile allows us to have a single source of truth
- Having them in the bakefile means we can more easily pass different values based on arch, instead of relying on shell-based ways of detecting the build architecture
- Using `ADD` directives instead of manual calls to `curl` is more idiomatic and also allows for better performance (works better with caching + parallel downloads)
- Add a couple more args and split things up (e.g. dda install script) to improve readability and uniformity of all the stages

refactor(linux): Use all tool-specific ARCH definitions

This makes it clearer which argument is used where, and avoids needing globally used `ARGs` (better for caching)

refactor(linux): Remove unused `build.env` files

fix(linux): Propery pass variables to LABEL instructions

* feat(bake): Add caching details to bakefile

* feat(ci): Use bake-based build logic in CI

* refactor(bake): Create common dummy stages for setting up local vs CI settings

* feat(linux): Make datadog-packages installation work locally and in CI

The problem solved here is that we were using a private repository for datadog-packages, so we need to authenticate to the repo in order to download it.
In CI, this is easy enough to do with the CI_JOB_TOKEN environment variable, but locally we need to use a "real" GitLab token.
This logic is now expressed flexibly in the bakefile.

* feat(bake): Specify dd-octo-sts version in bakefile

* fix(linux): Use dedicated stage for dd-octo-sts

The problem is that `--from` does not support variable interpolation, but declaring a stage and using it as a source for the next stage does
See https://forums.docker.com/t/how-can-i-expand-a-variable-within-a-copy-command-in-the-dockerfile/113538/7

* refactor: Remove useless stuff

* fix(linux): Add `MS_GOTOOLCHAIN_TELEMETRY_ENABLED=0` to final stage

* fix(linux): Move all Bazelisk logic to final stage

We're just downloading a file and symlinking it, so it's really not worth giving it its own stage

* fix(ci): Add 1h30 timeout to linux build jobs

* chore(bake): Remove unused `CI_DEFAULT_BRANCH` var and `linux-arm64-bazelisk` target

* fix(bake): Remove 8-character requirement from pipeline IDs

Gitlab CI IDs just count up and are global to the entire instance, so they will eventually become 9 characters.

* fix(bake): Use bazel `8.4.2`

That is the version currently defined in the agent repo